Make stream clients byte/string agnostic
Prior to this change clients were aware of the actual encoding used.
This info belongs to the client code (i.e. writing a csv vs a pickle)
not the client itself.

This change leaves the decision the code opening the stream not
hardcoding the instantion of the buffers in the clients