Hide strings/byte streams to S3 clients
S3/botocore doesn't really plays nicely with the duality string/byte
array native in python. This commit is an effort to:
1. Maintain the file mode outside of the client domain.
2. Encapsulate the responsability of having the creating the right stream type in
StreamClient related classes.

The chain of responsability of opening a url is:
URL.open -> handler: creates a reader/writer with of the right type ->
StreamClient, interfaces as a IO element (write/read) using -> client: acccess
the third party library elements.