Have bytearray to be our cannonical format
Prior to this change there was an issue in terms of giving support to
different clients in both string and bytearray, but we have decided to
move towards a pure binary approach. So all the clients (S3, ftp...).

This change wraps the stream readers and writers in a way that they hide
string streams from the clients, so someone opening `open(url, "wt")`
will be able to write stings but the client will recieve bytes.