Make db csv writer a StreamIO
Prior to this change we were missing some of the extra functionality
from StreamIO from the csv data writer. This, linked to some of the
latest changes in the core of data-io made the buffer to break.

This change fixes this issue and includes a functional test to make sure
this doesn't happen in the future