Delete protocols sanity check
Prior to this change there were some sanity checks done with protocols,
this was done in an early stage of the poject where the interaction with
the protocols was a bit experimental. They are being used everywhere now
and the tests are redundant and don't really apport any real benefit.